+++ src/pong/model/Paddle.java
@@ -2,35 +2,35 @@ package pong.model;
 
 public class Paddle extends MoveableItem {
 	
-	private double height;
-	private double width;
-	private double depth;
+	private float height;
+	private float width;
+	private float depth;
 
-	public Paddle(double xPos, double yPos, double zPos) {
+	public Paddle(float xPos, float yPos, float zPos) {
 		super(xPos, yPos, zPos);
 	}
 
-	public double getHeight() {
+	public float getHeight() {
 		return height;
 	}
 
-	public void setHeight(double height) {
+	public void setHeight(float height) {
 		this.height = height;
 	}
 
-	public double getWidth() {
+	public float getWidth() {
 		return width;
 	}
 
-	public void setWidth(double width) {
+	public void setWidth(float width) {
 		this.width = width;
 	}
 
-	public double getDepth() {
+	public float getDepth() {
 		return depth;
 	}
 
-	public void setDepth(double depth) {
+	public void setDepth(float depth) {
 		this.depth = depth;
 	}
